+++ src/datasources/alarms/components/editors/alarms-count/AlarmsCountQueryEditor.tsx
@@ -1,0 +1,14 @@
+import { AlarmsCountQuery } from "datasources/alarms/types/AlarmsCount.types";
+import React from "react";
+
+type Props = {
+  query: AlarmsCountQuery;
+};
+
+export function AlarmsCountQueryEditor(_: Props) {
+  return (
+    <>
+        <span>Placeholder for Alarms Count Query Editor</span>
+    </>
+  );
+}