+++ src/datasources/alarms/constants/defaultQueries.ts
@@ -1,0 +1,6 @@
+import { AlarmsCountQuery } from "../types/AlarmsCount.types";
+import { QueryType } from "../types/types";
+
+export const defaultAlarmsCountQuery: Omit<AlarmsCountQuery, 'refId'> = {
+    queryType: QueryType.AlarmsCount,
+};