+++ src/datasources/alarms/types/AlarmsCount.types.ts
@@ -1,0 +1,3 @@
+import { AlarmsQuery } from "./types";
+
+export interface AlarmsCountQuery extends AlarmsQuery {};