+++ src/datasources/alarms/types/types.ts
@@ -1,0 +1,9 @@
+import { DataQuery } from '@grafana/schema';
+
+export interface AlarmsQuery extends DataQuery {
+  queryType?: QueryType;
+}
+
+export enum QueryType {
+  AlarmsCount = 'AlarmsCount',
+}