+++ src/datasources/asset/types/types.ts
@@ -17,6 +17,7 @@ export interface AssetFeatureToggles {
   calibrationForecast: boolean;
   assetList: boolean;
   assetSummary: boolean;
+  locations: boolean;
 }
 
 export interface AssetDataSourceOptions extends DataSourceJsonData {
@@ -26,7 +27,8 @@ export interface AssetDataSourceOptions extends DataSourceJsonData {
 export const AssetFeatureTogglesDefaults: AssetFeatureToggles = {
   assetList: true,
   calibrationForecast: true,
-  assetSummary: true
+  assetSummary: true,
+  locations: false
 }
 
 export enum AssetQueryReturnType {